List Articles Implemented

+++ MB.Domain/ArticleAgg/Article.cs
@@ -5,20 +5,20 @@ namespace MB.Domain.ArticleAgg
 {
     public class Article
     {
-        public int Id { get;private set; }
+        public long Id { get;private set; }
         public string Title { get; private set; }
         public string ShortDescription { get; private set; }
         public string Content { get; private set; }
         public string Image { get; private set; }
         public bool IsDeleted { get; private set; }
         public DateTime CreationDate { get; private set; }
-        public int ArticleCategoryId { get; private set; }
+        public long ArticleCategoryId { get; private set; }
         public ArticleCategory ArticleCategory { get; private set; }
 
         protected Article()
         {
         }
-        public Article(string title, string shortDescription, string content, string image, int articleCategoryId)
+        public Article(string title, string shortDescription, string content, string image, long articleCategoryId)
         {
             Title = title;
             ShortDescription = shortDescription;
